Fixes bug with logging level by using the log level in the configuration as a filter
