+++ src/Google/Ads/GoogleAds/Lib/V1/GoogleAdsUnaryCallLogger.php
@@ -18,6 +18,7 @@
 namespace Google\Ads\GoogleAds\Lib\V1;
 
 use Google\ApiCore\ArrayTrait;
+use Monolog\Logger;
 use Psr\Log\LoggerInterface;
 use Psr\Log\LogLevel;
 
@@ -27,13 +28,14 @@ use Psr\Log\LogLevel;
 final class GoogleAdsUnaryCallLogger
 {
     use ArrayTrait;
-    use GoogleAdsLoggerTrait;
 
     private $logger;
     private $logLevel;
     private $endpoint;
     private $logMessageFormatter;
     private $context;
+    private static $logLevelList;
+    private static $logLevelNamesToNormalizedValues;
 
     /**
      * Constructs the Google Ads unary call logger with the specified PSR-3 logger interface.
@@ -56,6 +58,10 @@ final class GoogleAdsUnaryCallLogger
         $this->endpoint = $endpoint;
         $this->logMessageFormatter = $logMessageFormatter ?: new LogMessageFormatter();
         $this->context = $context;
+        if (!isset(self::$logLevelList) || !isset(self::$logLevelNamesToNormalizedValues)) {
+            self::$logLevelList = array_keys(Logger::getLevels());
+            self::$logLevelNamesToNormalizedValues = array_flip(self::$logLevelList);
+        }
     }
 
     /**
@@ -68,11 +74,14 @@ final class GoogleAdsUnaryCallLogger
         array $requestData,
         array $responseData
     ) {
-        $this->logger->log(
-            $this->getAppropriateLogLevel($responseData['status']->code),
-            $this->logMessageFormatter->formatSummary($requestData, $responseData, $this->endpoint),
-            $this->context
-        );
+        $level = $this->getAppropriateLogLevel($responseData['status']->code);
+        if ($this->isEnabled($level)) {
+            $this->logger->log(
+                $level,
+                $this->logMessageFormatter->formatSummary($requestData, $responseData, $this->endpoint),
+                $this->context
+            );
+        }
     }
 
     /**
@@ -85,26 +94,58 @@ final class GoogleAdsUnaryCallLogger
         array $requestData,
         array $responseData
     ) {
-        $this->logger->log(
-            // Logs details at one finer level than the summary.
-            self::getNextFinerLogLevel(
-                $this->getAppropriateLogLevel($responseData['status']->code)
-            ),
-            $this->logMessageFormatter->formatDetail($requestData, $responseData, $this->endpoint),
-            $this->context
+        // Logs details at one finer level than the summary.
+        $level = self::getNextFinerLogLevel(
+            $this->getAppropriateLogLevel($responseData['status']->code)
         );
+
+        if ($this->isEnabled($level)) {
+            $this->logger->log(
+                $level,
+                $this->logMessageFormatter->formatDetail($requestData, $responseData, $this->endpoint),
+                $this->context
+            );
+        }
     }
 
     /**
-     * Sets up the log level in case it's not specified by the user.
+     * Returns the appropriate log level depending on the response code.
      * For successful requests, use INFO. For failed requests, use WARNING.
+     * @return string the log level to use
      */
-    private function getAppropriateLogLevel($code)
+    private function getAppropriateLogLevel($code): string
     {
-        $logLevel = $this->logLevel;
-        if (is_null($logLevel)) {
-            $logLevel = $code === 0 ? LogLevel::INFO : LogLevel::WARNING;
+        return $code === 0 ? LogLevel::INFO : LogLevel::WARNING;
+    }
+
+    /**
+     * Returns true if $level is enabled, i.e. if the log level that this logger supports is
+     * less than or equal to $level.
+     * @return bool true if $level is enabled, false otherwise
+     */
+    private function isEnabled(string $level): bool
+    {
+        return self::$logLevelNamesToNormalizedValues[$this->logLevel] <=
+                self::$logLevelNamesToNormalizedValues[$level];
+    }
+
+    /**
+     * Returns the next finer PSR-3 log level for the specified log level.
+     *
+     * @param string $level the current log level
+     * @return string the level name
+     */
+    private static function getNextFinerLogLevel($level)
+    {
+        $currentLevel = self::$logLevelNamesToNormalizedValues[strtoupper($level)];
+        if (!isset($currentLevel)) {
+            throw new \InvalidArgumentException("The specified log level '$level' is invalid.");
         }
-        return $logLevel;
+        if ($currentLevel === 0) {
+            // DEBUG is the finest level, so returns itself instead.
+            return $level;
+        }
+
+        return self::$logLevelList[$currentLevel - 1];
     }
 }
